Adds spinner for clone operation
The "clone" operation is often long-running, and providing no user feedback
about a long-running process is bad user experience. A spinner, even though it
provides no additional information, reassures users that work is happening in
the background.

Spinner requires Halo but that could be built straight into the project
at the cost of additional complexity.