(#142) Add "forget" and "fo" alias to untrack
Untrack isn't common, as https://github.com/gitless-vcs/gitless/issues/142
points out. More common is "forget", which I think is an apt opposite to
"track". Adding forget as an alias seems like a reasonable compromise to me, as
it won't break anything might be a little clearer or easier to remember for
some people.